Move class HtmxSpringStandardExressionObjectFactory to thymeleaf and shorten the class name
The class is thymeleaf related and should not be part of the mvc package.
